// frontend/src/auto-api/models/FonInstructionDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type FonInstructionDto = {
    instructionId?: string;
    memberCode?: string;
    fonCode?: string;
    toMemberCode?: string;
    currency?: string;
    instructionDate?: string;
    valorDate?: string;
    count?: number;
    price?: number;
    operationType?: 'BUY' | 'SELL';
};